changed the models, shifted over the class
