+++ wingspan/tanalysis/views.py
@@ -1,12 +1,12 @@
 from django.shortcuts import render
 from django.http import HttpResponse
-from .models import Tweets
+from .models import TwitterCom
 
 # Create your views here.
 def index(request):
     try:
-        twit_launcher = Tweets()
-        tweets = twit_launcher.search_tweets("ruggs", 20)
+        twitter_launcher = TwitterCom()
+        tweets = twitter_launcher.search_tweets("bachelorette", 20)
     except:
         tweets = []
     html = "<html><body> %s </body></html>" % tweets
